Assignment 2/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
asgiref==3.8.1
chardet==5.2.0
Django==5.2
djangorestframework==3.16.0
pillow==11.1.0
reportlab==4.3.1
sqlparse==0.5.3
